Set default keepalive options for peer clients (#987)
The latest version of grpc now sets the default
client keepalive time to 10 seconds.  Prior to this,
if the timeout was set to 0, grpc actually used an
infinite timeout by default which effectively meant
no keepalives were ever sent.  Since it now sets the
default to 10 seconds and the server permitted timeout is
actual set to 20 seconds, we now see disconnects for
requests taking > 20 secs as the client is now violating
the policy.

Signed-off-by: Gari Singh <gari.r.singh@gmail.com>