mv keys/update_batches to ledger/internal/state
We have a bunch of composite keys, an update batch to
hold block's public, hashed, and private states, an update
batch to hold combined states to update the state database.

As these structs/methods are used throughout the ledger package,
we move these structs and related methods to
core/ledger/internal/state. Further, we increase the test
coverage and remove dead-code.

Signed-off-by: senthil <cendhu@gmail.com>