rm proto msg folder present in stateleveldb (#996)
In goleveldb, we store a KV pair of format {key, [value, version, metadata]}
for every state managed by the blockchain. As the [value, version, metadata] must
be passed as single []byte (as per goleveldb requirements), we use a proto to
serialize the combined values.

Currently, we have a separate package to have this proto. This is unjustifiable.
Hence, we move this proto to stateleveldb package itself and change the proto
message to DBValue from VersionedValueProto for better readability.

Signed-off-by: senthil <cendhu@gmail.com>