mv storageutil to ledger/internal and rm dead-code
The storageutil package provides API to
serialize and deserialize state metadata that
holds key-based endorsement policy. Note that the state
metadata is very generic and we can add any number of metadata
per key not just key-based endorsement policy.

The package name storageutil is confusing. Hence, we
rename it as statemetadata and also remove some dead-code.

Signed-off-by: senthil <cendhu@gmail.com>