replace Metadata in query with Limit and Bookmark
The range query receives metadata such as pagination limit.
The execute query at couchdb receives metadata such as the
pagination bookmark and pagination limit.
ValidateRangeMetadata() and ValidateQueryMetadata() validate
the received metadata.

However, the term metadata is confusing with the statemetadata
which stores the key-based endorsement. Hence, we replace
Metadata in range query API with limit and in the execute
query API with bookmark and limit. As we have removed the
metadata map, we also removed ValidateRangeMetadata()
and ValidateQueryMetadata() as they are not needed anymore.

Signed-off-by: senthil <cendhu@gmail.com>