mv rangequery option validation to ledger/internal
The range query receives metadata such as pagination limit.
ValidateRangeMetadata() validates the received metadata. As
the ValidateRangeMetadata() is present in statedb, it is imported
in both statecouchdb and stateleveldb packages. As we want these
packages to be self-sufficient and only import ledger/internal
packages, we move the ValidateRangeMetadata() to
ledger/interal/rangequery.

Further, the term metadata is confusing with the statemetadata
which stores the key-based endorsement. Hence, we replace
Metadata in range query API with options.

Signed-off-by: senthil <cendhu@gmail.com>