rm statedb package import in stateleveldb/statecouchdb
We have imported statedb within statelevel and statecouchdb
packages mainly to return the defined interface. However, the
both the stateleveldb and statecouchdb should not be returning
the interface that they implement. The interface should be
defined and used only at the consumer side. Hence, this PR
removes all statedb package imports in both of these packages.

Signed-off-by: senthil <cendhu@gmail.com>