Properly handle malformed gossip envelopes
If a malformed envelope is read from the stream, an error is propagated
synchronously up the stack.

However, the envelope is unmarshaled into a nil
message which is also propagated further up the stack asynchronously.

Under very rare circumstances, the error is picked up later than
the message, and a nil pointer panic occurs.

This patch fixes this by returning early in case of an error.

Change-Id: Ia17767ec2483d83d5fa4e7e22514c539232108a8
Signed-off-by: yacovm <yacovm@il.ibm.com>