[FAB-17716] Fix test flake due to too many requests for /protos.Deliver
Currently peer.limits.concurrency.deliverService value is set to 1
in IT core_template.go. This value is too low and may cause test flakes
because some tests have multiple goroutines sending requests to DeliverService.

This PR sets a higher value for all tests in IT. Meanwhile, it overwrites
the value to 1 in pvtdata_test.go where maximum 1 goroutine sends requests
to DeliverService.

Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>