Improve chaincode validation error message
It was not clear if chaincode lifecycle error messages were coming
from some problem in peer or from validation on the passed chaincode definition.

This change makes it clear which errors are a result of validation
on the passed chaincode definition, so that users can know
that their definition needs to be fixed.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>