CouchDB UT flakes -- index creation
As per CI statistics, TestHandleChaincodeDeployErroneousIndexFile has
failed 152 times across 173 pipeline runs in the last 14 days.

The potential reason for this error is that the CouchDB creates an
index in the background and responds to the REST API immediately.
Currently, we wait only for 100 milliseconds to allow CouchDB
to build the index before issuing the query. This 100 msecs sleep
in inadequate even for some local test environment used for the
development. Hence, this PR increases the sleep time to 2 seconds
and allow ample time for CouchDB to build the index before issuing
the query.

Signed-off-by: senthil <cendhu@gmail.com>