Add 1ms delay to send loop in TestCommSend
The existing implementation of TestCommSend sends messages in a tight
loop. Because of the production code implementation (see FAB-17736) this
results in thousands of go routines.

This change slows the creation of go routines to approximately 2000 per
second by throttling the sends to 1 per ms per sender.

This also replaces the use of a ticker with an timer in the one-shot
timeout processing and DRYs up the send loop.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>