rm unnecessary mock from txmgr/validator
In the validator, we already create a mock for TxMgr interface
but use the mock created at txmgr package. Hence, we make the
validator to use the mock that present locally.

Further, the validator does not need to create mock for all APIs
in the TxMgr interface as it only uses NewTxSimulator(). Hence,
we define a local interface.

Signed-off-by: senthil <cendhu@gmail.com>