Remove concurrent session creation from raft suite
The session creation commands in nwo.Network were to be used with ginkgo and
were not structured for concurrent access. However, one integration test suite
that pushes the boundaries and starts sessions in parallel using go
routines in the test suites.

Removing the go routine prevents a concurrent map access in the network,
simplifies the test, and does not seem to impact the execution time of
the suite. (Before and after execution were both over 9 minutes and the
version without concurrency was 4 seconds faster locally.)

Change-Id: I443f02958ad61e800ff70e8dc6ba622ee14f8461
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>