make couchdb index processor deterministic
We need to satisfy two requirements while processing the index files.
R1: all valid indexes should be processed.
R2: the order of index creation must be the same in all peers. For example, if user
passes two index files with the same index name but different index fields and we
process these files in different orders in different peers, each peer would
have different indexes (as one index definion could replace another if the index names
are the same).

To satisfy R1, we log the error and continue to process the next index file.
To satisfy R2, we sort the indexFilesData map based on the filenames and process
each index as per the sorted order.

Signed-off-by: senthil <cendhu@gmail.com>