[LEDGER] merge queryHelper with the queryExecutor (#1097)
In lockbasedtxmgr, we have a lockBasedQueryExecutor and a
lockBasedTxSimulator struct. The lockBasedTxSimulator implements
mainly SetXXX() APIs while lockBasedQueryExecutor implements GetXXX()
APIs. As the transaction simulation also involves query, the
lockBasedTxSimulator encapsulate lockBasedQueryExecutor. Only in
the lockBasedTxSimulator, we collect readSet while using GetXXX()
and writeSet while using SetXXX(). However, when we use
lockBasedQueryExecutor alone, we don't collect readSet.

Currently, we have an unnecessary indirection with QueryHelper
and it makes code readability a bit hard. Hence, we merge
the QueryHelper with the lockBasedQueryExecutor itself
and do some renaming of files/structs too.

Signed-off-by: senthil <cendhu@gmail.com>