Return Error From PKCS11 CreateSession
The current implementation of BCCSP performs a fatal
logging function in the event it cannot open a session
with the HSM. This is fine during startup, but the
createSession function is called during the standard
code flow, i.e., when the cryptoServiceProviders sessions
pool is already completely in use (by default 10 sessions)
a new session is created. So it is possible while during
running for the fatal log to occur when an expected error
occurs, i.e., the HSM has reached its maximum session handles.
Instead of issuing a fatal log call, we should return the error
as it's probable the error is recoverable, or outside the control
of Fabric itself (is a problem with the HSM).

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>