Improve lifecycle error message for committed sequence
If you attempt to approveformyorg again for a committed chaincode definition sequence,
for example to update chaincode package on a peer, but you pass the incorrect
chaincode definition parameters, the error message was not clear.

This change updates the error message to something that the admin can likely
troubleshoot themselves.

Old message:
Error: proposal failed with status: 500 - failed to invoke backing implementation
of 'ApproveChaincodeDefinitionForMyOrg': attempted to define the current sequence
(2) for namespace marbles5, but: Version '3' != '4'

New message:
Error: proposal failed with status: 500 - failed to invoke backing implementation
of 'ApproveChaincodeDefinitionForMyOrg': attempted to redefine the current
committed sequence (3) for namespace marbles5 with different parameters:
expected Version '3' does not match passed Version '4'

Signed-off-by: David Enyeart <enyeart@us.ibm.com>