Stop modifying curve held in crypto package var
When TestSignECDSABadParameter modified the N parameter of the generated
private key, it was actually modifying the copy of the curve maintained
by the standard library. Since it modified the standard library, it
would restore the original value out of a defer.

Simply copying the curve would prevent any modification to the package
level state from the standard library and remove the need to save and
restore the original N value of the curve.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>