check len. of indexDir before accessing
While processing the couchdb index files, we retrieve and store
all directories in the index filepath into an array in order to
find whether it is a chaincode index or collection index. Currently,
we access the array without checking the number of directories
in the path assuming that the user would pass the correct path.
When the user passes a wrong path, the current array access would
result in a peer panic. Hence, we add a check on the length of
the array before accessing it.

FAB-17771

Signed-off-by: senthil <cendhu@gmail.com>