[LEDGER] ensure peer panic does not occur when the index path is wrong (#1181)
* check len. of indexDir before accessing

While processing the couchdb index files, we retrieve and store
all directories in the index filepath into an array in order to
find whether it is a chaincode index or collection index. Currently,
we access the array without checking the number of directories
in the path assuming that the user would pass the correct path.
When the user passes a wrong path, the current array access would
result in a peer panic. Hence, we add a check on the length of
the array before accessing it.

FAB-17771

Signed-off-by: senthil <cendhu@gmail.com>

* rm privacyenabledstate_test pkg and gomega

To add test for the getIndexInfo(), we need to use the
same package name for the test files too. Hence, we replace
the privacyenabledstate_test with privacyenabledstate. Further,
we remove the usage of gomega in the UT.

Signed-off-by: senthil <cendhu@gmail.com>

* add test for getIndexInfo

Signed-off-by: senthil <cendhu@gmail.com>