[FAB-17433] Add _lifecycle function for querying approved chaincode
Currently, Fabric peer does not provide a function to query the details of
approved chaincode definitions.

On the other hand, Fabric admins will need to confirm/use the information
after approval for their operations.

This patch adds a function for querying the details of the chaincode
definition approved by the organization itself to `_lifecycle`,
so that Fabric peer can provide a function to query them.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>