test for commit hash computation's backward compatibility
We might change the code that computes the block commit hash
during refactoring or to improve the performance from time to time.
When changing the hash computation code, we need to ensure that
we are not changing the way commit hash has been computed in previous
releases. Hence, we add a test to ensure that commit hash
computation is always compatible with the first code added in v1.1

FAB-17878

Signed-off-by: senthil <cendhu@gmail.com>