Reset golang.org/x/sys to commit ef0ce1748380
A breaking change was introduced to the golang.org/x/sys/windows package
API that causes compilation failures in docker packages indirectly
consumed by Fabric. While the docker code has been updated to
accommodate the broken API, the change is not available in any release.

Instead of using the development branch of docker, this change points
the golang.org/x/sys dependency to the commit prior to the API change.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>