[FAB-17886] Fix default value of `sequence` parameter to empty
Currently, default value of `sequence` parameter for `peer chaincode lifecycle`
commands is set to `1`.

However, the default value should be modified to `0` that means empty for
the following several reasons.

- According to `Validate()` in `approveformyorg.go`, this function is
implemented assuming that `sequence` is `0`. This is inconsistent with
the default value `1`.

- The current default value `1` enables Fabric users `approveformyorg` without
specifying the sequence number. Not forcing the specification of a sequence
number may cause operation mistakes (e.g, updating the package-id for
sequence 1 by mistake) when running `approveformyorg` in the second and
subsequent times.

- The default value is preferably `0` to reuse the same parameter `sequence`
in CLI commands to query the details of approved chaincode definitions.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>