[FAB-17436] Add documents for `queryapproved` command
Currently, Fabric peer does not provide a function to query the details of
approved chaincode definitions.

On the other hand, Fabric admins will need to confirm/use the information
after approval for their operations.

Other patches add `queryapproved` command for querying the details of
the approved chaincode definition to Fabric peer CLI
(Refer to FAB-17433 and FAB-17434).

This patch adds some documents on how to use the CLI command.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>