Pass HashProvider to ledger instead of Hasher
So far ledger used hash function for computing hash on
small amount of data at a time that was kept in memory and hence
Hasher (from bccsp) was passed as a dependency. However, for computing
the hash of the data streaming while exporting the data for a snapshot,
the access to to Hash object is needed.

This commit makes the change in this dependency. However, no change is
needed on the upper layer as the already passed down dependency already
implements the desired function.

Signed-off-by: manish <manish.sethi@gmail.com>