Silence usage and errors during cli tests
The tests exercise a number of commands that result in usage and error
messages going to the console. Since the tests are asserting against the
errors and not the output, we can silence the output to prevent noise
during execution.

For logging, redirect the output to the GinkgoWriter to suppress it when
not in verbose mode yet ensure it's present when things fail.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>