Simplify extractOrdererConsensusType
The error that was returned is unnecessary as the function is only used
and test and can therefore perform the assertion.

The "updated" value was rarely used and getting flagged as an
"ineffectual assignment" in linting. Since all the function was doing
was cloning the input, the clone was moved to the places where it was
necessary.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>