mv dbvalue proto to statedb
The DBValue proto encloses bytes of version, value, and metadata.
This proto message has multiple use:
(1) store the version, value, and metadata as a bytes slice.
(2) return the value associated with a key as a bytes slice.

While (1) is used only in the stateleveldb, (2) would be used
when the key-value pairs are exported from both leveldb and
statecouchdb.

We have statedb.VersioneKV that looks similar to DBValue proto
except that version is not represented as bytes. In a separate PR,
we can remove statedb.VersionedKV and use DBValue at all places.

Signed-off-by: senthil <cendhu@gmail.com>