fix infinite loop during full range query (#1347)
We use a single scanner for achieving both paginated and
non-paginated range query.

We have internalQueryLimit and pageSize. For each
_all_docs?startKey="XXX"&endKey="YYY" REST API call to
CouchDB, we fetch atmost internalQueryLimit number of
records only by appending limit=internalQueryLimit.

When the requested pageSize is higher than the internalQueryLimit
or the total number of records present in the given range
is higher than the internalQueryLimit, iterator would execute the
query again once the records fetched in the first cycle is consumed
and so on. In order to do that, after an execution of the REST API in each
cycle, it updates the initially passed startKey to the nextStartKey.
If there is no nextStartKey, it is set to the endKey.

Currently, when the nextStartKey and the endKey are the same, we still
run one REST API call which is actually not needed as we always set
inclusive_end=false. However, this causes infinite loop in a particular
case. When we want to retrieve all the records, we would pass an empty
string as the startKey and the endKey. When the startKey is an empty
string, the REST API call would become _all_docs?endKey="YYY". When
both are empty, it would become _all_docs

Given that we set startKey to endKey when there is no nextStartKey and
still execute one REST API call, it gets into infinite loop by fetching
all records again and again.

We avovid this infinite loop by setting scanner.exhausted = true whe
the startKey and endKey become the same when there is no nextStartKey

Signed-off-by: senthil <cendhu@gmail.com>