FAB-17961 Ch.Part.API: Refactor onboarding out of server
Part 1 of 3 commits: here the onboarding code is refactored out
of the common/server package in to an (almost) leaf package that
depends mainly on cluster. This is done in order to make it server
main() cleaner and to allow the multichannel as well as other
packages to reuse the onboarding code (break the import cycle).

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: Ia8bea790132d3ee6e182a0a5e0dc9b0da40c731f