FAB-17961 Ch.Part.API: Refactor main to setup cluster when no system channel
Part 2 of 3 commits: here the server main.go is refactored in order to make
it clearer and simpler, and setup cluster when there is no system channel
and no bootstrap block. We assume that is this case the network is of
cluster type, i.e. etcdraft, and therefore setup everything we can to
allow a correct initialization of the cluster.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I26352d1e8a04082fcc7d1e96b8169990d7818c05