Bump Go and Alpine Version
Bump Go to 1.14.4. This is also the first Go release
to switch to alpine 3.12 with 1.14.3 being the last
to support 3.11. Performing this update now will allow
us to cross the alpine minor version barrier prior
to our LTS release of 2.2 to give us time to test
the stability of the release.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>