FAB-17985 Registrar newChain not to copy chains map
Currently the Registrar.newChain() copies the chains map to a new one,
supposedly to allow concurrent reads on the map while a chain is created.
This is redundant because the newChain takes rwmutex.Lock(), which blocks reads.

It is enough to just create the chain and insert it to the map.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I61ca191a4d0529210163b0fa4711509ac6669fa2