[FAB-17992] Add remove channel func to blockstorage provider
Add a Remove function to block storage provider in oder to
remove ledger data for a channel. It creates a temporary file
to indicate the channel is to be removed and start a goroutine
to remove channel ledger data in background. If remove fails or
orderer is stopped before remove is done, upon ledger restart,
it checks the existence of the temporary file and complete remove
as needed.

Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>