FAB-18035 Ch.Part.API: consenter JoinChain creates a follower.Chain
Implement the JoinChain() method in the consenter, and create a
skeleton implementation of a follower.Chain.

Create a cluster.BlockPuller that is initiated from the join-block.

Create a SelfMembershipPredicate for the follower.Chain.

The JoinChain method tries to be strict and checj the inputs before
chain creation, in order to minimize the risk errors after the
chain is started.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I94043a66c73488e32ac2a49606886e1f94be5cf0