Ch. Part. API: etcdraft consenter w/o system channel (#1376)
* FAB-17961 Ch.Part.API: Refactor onboarding out of server

Part 1 of 3 commits: here the onboarding code is refactored out
of the common/server package in to an (almost) leaf package that
depends mainly on cluster. This is done in order to make it server
main() cleaner and to allow the multichannel as well as other
packages to reuse the onboarding code (break the import cycle).

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: Ia8bea790132d3ee6e182a0a5e0dc9b0da40c731f

* FAB-17961 Ch.Part.API: Refactor main to setup cluster when no system channel

Part 2 of 3 commits: here the server main.go is refactored in order to make
it clearer and simpler, and setup cluster when there is no system channel
and no bootstrap block. We assume that is this case the network is of
cluster type, i.e. etcdraft, and therefore setup everything we can to
allow a correct initialization of the cluster.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I26352d1e8a04082fcc7d1e96b8169990d7818c05

* FAB-17961 Ch.Part.API: Create an etcdraft consenter when there is no system channel

Part 3 of 3 commits: here the intialization of an etcdraft consenter
in main.go is performed even when there is no system channel and
no bootstrap block. The orderer can now start without channels at
all and have an initialized consenter,  ready for the creation of
new channels.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I20caf39e33395447b8d1a5c8f606948bed2ccb94

* Review comments: tune log messages

- Tune log messages
- Check the registrar has an etcdraft consenter when it starts without channels

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I0214949251c14efa68744024b5d655a23aa30ffc

* Review comments: simplify main

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I237d0cfa018d5689180bcb137ac12d2c5daf3ffc