Remove call to couchdb API _ensure_full_commit
This PR removes the invocation to the couchdb API "_ensure_full_commit".
We use "delayed_commits=true" in couchdb config for better performance
and then at the block commit time, we invoke "_ensure_full_commit".
This arrangement was under the assumption that the "_ensure_full_commit"
fsyncs the data as per the couchdb docs https://docs.couchdb.org/en/2.3.1/api/database/compact.html#db-ensure-full-commit.
However, after talking to couchdb experts, we figured that this was true
only till couchdb 1.x. Since then, this is more or less a no-op call as
far as `fsync` is concerned.

Signed-off-by: manish <manish.sethi@gmail.com>